+++ packages/@expo/cli/src/start/server/webpack/WebpackBundlerDevServer.ts
@@ -4,6 +4,7 @@ import type { Application } from 'express';
 import fs from 'fs';
 import http from 'http';
 import * as path from 'path';
+import ProgressBar from 'progress';
 import resolveFrom from 'resolve-from';
 import type webpack from 'webpack';
 import type WebpackDevServer from 'webpack-dev-server';
@@ -13,8 +14,10 @@ import { env } from '../../../utils/env';
 import { CommandError } from '../../../utils/errors';
 import { getIpAddress } from '../../../utils/ip';
 import { choosePortAsync } from '../../../utils/port';
+import { createProgressBar } from '../../../utils/progress';
 import { ensureDotExpoProjectDirectoryInitialized } from '../../project/dotExpo';
 import { BundlerDevServer, BundlerStartOptions, DevServerInstance } from '../BundlerDevServer';
+import { compileAsync } from './compile';
 import {
   importExpoWebpackConfigFromProject,
   importWebpackDevServerFromProject,
@@ -165,6 +168,59 @@ export class WebpackBundlerDevServer extends BundlerDevServer {
     }
   }
 
+  async bundleAsync({ mode, clear }: { mode: 'development' | 'production'; clear: boolean }) {
+    // Do this first to fail faster.
+    const webpack = importWebpackFromProject(this.projectRoot);
+
+    if (clear) {
+      await this.clearWebProjectCacheAsync(this.projectRoot, mode);
+    }
+
+    const config = await this.loadConfigAsync({
+      isImageEditingEnabled: true,
+      mode,
+    });
+
+    if (!config.plugins) {
+      config.plugins = [];
+    }
+
+    let bar: ProgressBar | null = null;
+
+    // NOTE(EvanBacon): Prevent breaking in CI.
+    if (process.stderr.clearLine != null) {
+      bar = createProgressBar(chalk`{bold Web} Bundling Javascript [:bar] :percent`, {
+        width: 64,
+        total: 100,
+        clear: true,
+        complete: '=',
+        incomplete: ' ',
+      });
+
+      // Add a progress bar to the webpack logger.
+      config.plugins.push(
+        new webpack.ProgressPlugin((percent: number) => {
+          bar?.update(percent);
+          if (percent === 1) {
+            bar?.terminate();
+          }
+        })
+      );
+    }
+
+    // Create a webpack compiler that is configured with custom messages.
+    const compiler = webpack(config);
+
+    try {
+      await compileAsync(compiler);
+    } catch (error: any) {
+      Log.error(chalk.red('Failed to compile'));
+      throw error;
+    } finally {
+      bar?.terminate();
+    }
+  }
+
   protected async startImplementationAsync(
     options: BundlerStartOptions
   ): Promise<DevServerInstance> {
@@ -189,7 +245,7 @@ export class WebpackBundlerDevServer extends BundlerDevServer {
     Log.debug('Starting webpack on port: ' + port);
 
     if (resetDevServer) {
-      await clearWebProjectCacheAsync(this.projectRoot, mode);
+      await this.clearWebProjectCacheAsync(this.projectRoot, mode);
     }
 
     if (https) {
@@ -292,7 +348,7 @@ export class WebpackBundlerDevServer extends BundlerDevServer {
   }
 
   async loadConfigAsync(
-    options: BundlerStartOptions,
+    options: Pick<BundlerStartOptions, 'mode' | 'isImageEditingEnabled' | 'https'>,
     argv?: string[]
   ): Promise<WebpackConfiguration> {
     // let bar: ProgressBar | null = null;
@@ -330,6 +386,21 @@ export class WebpackBundlerDevServer extends BundlerDevServer {
   protected getConfigModuleIds(): string[] {
     return ['./webpack.config.js'];
   }
+
+  protected async clearWebProjectCacheAsync(
+    projectRoot: string,
+    mode: string = 'development'
+  ): Promise<void> {
+    Log.log(chalk.dim(`Clearing Webpack ${mode} cache directory...`));
+
+    const dir = await ensureDotExpoProjectDirectoryInitialized(projectRoot);
+    const cacheFolder = path.join(dir, 'web/cache', mode);
+    try {
+      await fs.promises.rm(cacheFolder, { recursive: true, force: true });
+    } catch (error: any) {
+      Log.error(`Could not clear ${mode} web cache directory: ${error.message}`);
+    }
+  }
 }
 
 function setMode(mode: 'development' | 'production' | 'test' | 'none'): void {
@@ -339,19 +410,4 @@ function setMode(mode: 'development' | 'production' | 'test' | 'none'): void {
 
 export function getProjectWebpackConfigFilePath(projectRoot: string) {
   return resolveFrom.silent(projectRoot, './webpack.config.js');
-}
-
-async function clearWebProjectCacheAsync(
-  projectRoot: string,
-  mode: string = 'development'
-): Promise<void> {
-  Log.log(chalk.dim(`Clearing Webpack ${mode} cache directory...`));
-
-  const dir = await ensureDotExpoProjectDirectoryInitialized(projectRoot);
-  const cacheFolder = path.join(dir, 'web/cache', mode);
-  try {
-    await fs.promises.rm(cacheFolder, { recursive: true, force: true });
-  } catch (error: any) {
-    Log.error(`Could not clear ${mode} web cache directory: ${error.message}`);
-  }
 }